chore: add wallet actions component
