+++ apps/web/src/components/WalletModalV2/WalletModal.tsx
@@ -17,7 +17,6 @@ import { RecentTransactions } from 'components/App/Transactions/TransactionsModa
 import { useTheme } from '@pancakeswap/hooks'
 import { useMenuTab, WalletView } from 'components/Menu/UserMenu/providers/MenuTabProvider'
 import { TabsComponent } from 'components/Menu/UserMenu/WalletModal'
-import { ASSET_CDN } from 'config/constants/endpoints'
 import { useAddressBalance } from 'hooks/useAddressBalance'
 import { useRouter } from 'next/router'
 import React, { useCallback, useMemo, useState } from 'react'
@@ -44,6 +43,7 @@ import { CopyAddress } from './WalletCopyButton'
 import { useWalletModalV2ViewState } from './WalletModalV2ViewStateProvider'
 import ReceiveOptionsView from './ReceiveOptionsView'
 import { ReceiveContent } from './ReceiveModal'
+import EmptyWalletActions from './EmptyWalletActions'
 
 interface WalletModalProps {
   isOpen: boolean
@@ -94,16 +94,6 @@ const DisconnectButton = styled(Button)`
   &:hover {
     opacity: 0.8;
   }
-`
-
-const OptionBox = styled(Box)`
-  background: ${({ theme }) => theme.colors.input};
-  border-radius: 24px;
-  padding: 16px;
-  width: 45%;
-  border: 1px solid ${({ theme }) => (theme.isDark ? '#55496E' : '#D7CAEC')};
-  text-align: center;
-  cursor: pointer;
 `
 
 const WalletModal: React.FC<WalletModalProps> = ({ account, onDismiss, isOpen, onReceiveClick, onDisconnect }) => {
@@ -345,67 +335,14 @@ export const WalletContent = ({
         <>
           {noAssets ? (
             chainId === NonEVMChainId.SOLANA ? (
-              <Text color="textSubtle" textAlign="center" mb="16px">
-                {t('This wallet looks new. Does not have any assets.')}
-              </Text>
+              <EmptyWalletActions
+                onDismiss={onDismiss}
+                onReceiveClick={onReceiveClick}
+                setViewState={setViewState}
+                description={t('This wallet looks new. Does not have any assets.')}
+              />
             ) : (
-              <Box padding="8px 16px">
-                <Text color="textSubtle" textAlign="center" mb="16px">
-                  {t('This wallet looks new — choose an option below to add crypto and start trading')}
-                </Text>
-                <FlexGap gap="16px" justifyContent="center" flexWrap="wrap">
-                  <OptionBox
-                    onClick={() => {
-                      router.push('/buy-crypto')
-                      onDismiss()
-                    }}
-                  >
-                    <Box mb="16px" mx="auto" width="60px" height="60px">
-                      <img src={`${ASSET_CDN}/web/landing/trade-buy-crypto.png`} width="60px" alt="Buy Crypto" />
-                    </Box>
-                    <Text bold color="secondary" fontSize="16px" mb="8px">
-                      {t('Buy')}
-                    </Text>
-                    <Text fontSize="14px" color="textSubtle">
-                      {t('Purchase with credit card, Apple Pay, or Google Pay.')}
-                    </Text>
-                  </OptionBox>
-                  <OptionBox
-                    onClick={() => {
-                      if (isMobile) {
-                        onReceiveClick()
-                        onDismiss()
-                      } else {
-                        setViewState(ViewState.RECEIVE_OPTIONS)
-                      }
-                    }}
-                  >
-                    <Box mb="16px" mx="auto" width="60px" height="60px">
-                      <img src={`${ASSET_CDN}/web/landing/earn-fixed-staking.png`} width="60px" alt="Receive Crypto" />
-                    </Box>
-                    <Text bold color="secondary" fontSize="16px" mb="8px" ml="8px">
-                      {t('Receive')}
-                    </Text>
-                    <Text fontSize="14px" color="textSubtle">
-                      {t('Receive crypto from another wallet.')}
-                    </Text>
-                  </OptionBox>
-                </FlexGap>
-                <FlexGap
-                  justifyContent="center"
-                  alignItems="center"
-                  mt="24px"
-                  onClick={() => {
-                    router.push('/bridge')
-                  }}
-                  style={{ cursor: 'pointer' }}
-                >
-                  <Text bold color="primary" fontSize="16px">
-                    {t('Bridge Crypto')}
-                  </Text>
-                  <ArrowForwardIcon color="primary" />
-                </FlexGap>
-              </Box>
+              <EmptyWalletActions onDismiss={onDismiss} onReceiveClick={onReceiveClick} setViewState={setViewState} />
             )
           ) : view === WalletView.GIFTS ? null : (
             <ActionButtonsContainer>
